@@ -5,7 +5,7 @@ import (
 )
 
 func TestCheckAccessNoFail(t *testing.T) {
-	c := createStubHTTPClient(t, "check-access-no-fail-req.xml", "check-access-no-fail-res.xml")
+	c := createStubHTTPClientSingle(t, "check-access-no-fail")
 
 	res, err := c.CheckAccessNoFail("", "", "")
 	FailOnError(t, err)
@@ -15,7 +15,7 @@ func TestCheckAccessNoFail(t *testing.T) {
 }
 
 func TestCheckAccess(t *testing.T) {
-	c := createStubHTTPClient(t, "check-access-req.xml", "check-access-res.xml")
+	c := createStubHTTPClientSingle(t, "check-access")
 
 	res, err := c.CheckAccess("", "", "")
 	FailOnError(t, err)
@@ -25,7 +25,7 @@ func TestCheckAccess(t *testing.T) {
 }
 
 func TestGetAuthnModuleName(t *testing.T) {
-	c := createStubHTTPClient(t, "get-authn-module-name-req.xml", "get-authn-module-name-res.xml")
+	c := createStubHTTPClientSingle(t, "get-authn-module-name")
 	var expected = "authentication.configfile"
 
 	res, err := c.GetAuthnModuleName()
@@ -36,7 +36,7 @@ func TestGetAuthnModuleName(t *testing.T) {
 }
 
 func TestLogin(t *testing.T) {
-	c := createStubHTTPClient(t, "login-req.xml", "login-res.xml")
+	c := createStubHTTPClientSingle(t, "login")
 	ok, err := c.Login()
 	FailOnError(t, err)
 
@@ -51,7 +51,7 @@ func TestLogin(t *testing.T) {
 }
 
 func TestLoginWithError(t *testing.T) {
-	c := createStubHTTPClient(t, "login-req.xml", "login-res-err.xml")
+	c := createStubHTTPClientSingle(t, "login-err")
 	expected := `Fault(1): <class 'cobbler.cexceptions.CX'>:'login failed (cobbler)'`
 
 	ok, err := c.Login()
@@ -59,13 +59,13 @@ func TestLoginWithError(t *testing.T) {
 		t.Errorf("false expected; got true")
 	}
 
-	if err.Error() != expected {
+	if err == nil || err.Error() != expected {
 		t.Errorf("%s expected; got %s", expected, err)
 	}
 }
 
 func TestLogout(t *testing.T) {
-	c := createStubHTTPClient(t, "logout-req.xml", "logout-res.xml")
+	c := createStubHTTPClientSingle(t, "logout")
 	var expected = false
 
 	res, err := c.Logout()
@@ -76,7 +76,7 @@ func TestLogout(t *testing.T) {
 }
 
 func TestTokenCheck(t *testing.T) {
-	c := createStubHTTPClient(t, "token-check-req.xml", "token-check-res.xml")
+	c := createStubHTTPClientSingle(t, "token-check")
 	var expected = false
 
 	res, err := c.TokenCheck("my_fake_token")
@@ -87,7 +87,7 @@ func TestTokenCheck(t *testing.T) {
 }
 
 func TestGetUserFromToken(t *testing.T) {
-	c := createStubHTTPClient(t, "get-user-from-token-req.xml", "get-user-from-token-res.xml")
+	c := createStubHTTPClientSingle(t, "get-user-from-token")
 	var expected = "testuser"
 
 	res, err := c.GetUserFromToken("securetoken99")
