@@ -6,7 +6,7 @@ import (
 )
 
 func TestGetTaskStatus(t *testing.T) {
-	c := createStubHTTPClient(t, "get-task-status-req.xml", "get-task-status-res.xml")
+	c := createStubHTTPClientSingle(t, "get-task-status")
 	expectedResult := CobblerEvent{
 		id:        "2022-09-30_200403_Updating Signatures_8f2b3c1626fb4b158636059b31242ee6",
 		statetime: 1664568243.5196018,
@@ -23,14 +23,14 @@ func TestGetTaskStatus(t *testing.T) {
 }
 
 func TestGetEvents(t *testing.T) {
-	c := createStubHTTPClient(t, "get-events-req.xml", "get-events-res.xml")
+	c := createStubHTTPClientSingle(t, "get-events")
 
 	_, err := c.GetEvents("")
 	FailOnError(t, err)
 }
 
 func TestGetEventLog(t *testing.T) {
-	c := createStubHTTPClient(t, "get-event-log-req.xml", "get-event-log-res.xml")
+	c := createStubHTTPClientSingle(t, "get-event-log")
 
 	_, err := c.GetEventLog("2022-09-30_145124_Sync_2cabdc4eddfa4731b45f145d7b625e29")
 	FailOnError(t, err)
