@@ -22,7 +22,7 @@ import (
 )
 
 func TestGetFiles(t *testing.T) {
-	c := createStubHTTPClient(t, "get-files-req.xml", "get-files-res.xml")
+	c := createStubHTTPClientSingle(t, "get-files")
 	files, err := c.GetFiles()
 	FailOnError(t, err)
 
@@ -32,7 +32,7 @@ func TestGetFiles(t *testing.T) {
 }
 
 func TestGetFile(t *testing.T) {
-	c := createStubHTTPClient(t, "get-file-req.xml", "get-file-res.xml")
+	c := createStubHTTPClientSingle(t, "get-file")
 	file, err := c.GetFile("testfile")
 	FailOnError(t, err)
 
@@ -42,19 +42,19 @@ func TestGetFile(t *testing.T) {
 }
 
 func TestDeleteFile(t *testing.T) {
-	c := createStubHTTPClient(t, "delete-file-req.xml", "delete-file-res.xml")
+	c := createStubHTTPClientSingle(t, "delete-file")
 	err := c.DeleteFile("test")
 	FailOnError(t, err)
 }
 
 func TestDeleteFileRecursive(t *testing.T) {
-	c := createStubHTTPClient(t, "delete-file-req.xml", "delete-file-res.xml")
+	c := createStubHTTPClientSingle(t, "delete-file")
 	err := c.DeleteFileRecursive("test", false)
 	FailOnError(t, err)
 }
 
 func TestListFileNames(t *testing.T) {
-	c := createStubHTTPClient(t, "get-item-names-file-req.xml", "get-item-names-file-res.xml")
+	c := createStubHTTPClientSingle(t, "get-item-names-file")
 	files, err := c.ListFileNames()
 	FailOnError(t, err)
 
@@ -64,7 +64,7 @@ func TestListFileNames(t *testing.T) {
 }
 
 func TestGetFilesSince(t *testing.T) {
-	c := createStubHTTPClient(t, "get-files-since-req.xml", "get-files-since-res.xml")
+	c := createStubHTTPClientSingle(t, "get-files-since")
 	files, err := c.GetFilesSince(time.Date(1970, 1, 1, 0, 0, 0, 0, time.UTC))
 	FailOnError(t, err)
 
@@ -74,7 +74,7 @@ func TestGetFilesSince(t *testing.T) {
 }
 
 func TestFindFile(t *testing.T) {
-	c := createStubHTTPClient(t, "find-file-req.xml", "find-file-res.xml")
+	c := createStubHTTPClientSingle(t, "find-file")
 	criteria := make(map[string]interface{}, 1)
 	criteria["name"] = "testfile"
 	files, err := c.FindFile(criteria)
@@ -86,7 +86,7 @@ func TestFindFile(t *testing.T) {
 }
 
 func TestFindFileNames(t *testing.T) {
-	c := createStubHTTPClient(t, "find-file-names-req.xml", "find-file-names-res.xml")
+	c := createStubHTTPClientSingle(t, "find-file-names")
 	criteria := make(map[string]interface{}, 1)
 	criteria["name"] = "testfile"
 	files, err := c.FindFileNames(criteria)
@@ -98,25 +98,25 @@ func TestFindFileNames(t *testing.T) {
 }
 
 func TestSaveFile(t *testing.T) {
-	c := createStubHTTPClient(t, "save-file-req.xml", "save-file-res.xml")
+	c := createStubHTTPClientSingle(t, "save-file")
 	err := c.SaveFile("file::testfile", "bypass")
 	FailOnError(t, err)
 }
 
 func TestCopyFile(t *testing.T) {
-	c := createStubHTTPClient(t, "copy-file-req.xml", "copy-file-res.xml")
+	c := createStubHTTPClientSingle(t, "copy-file")
 	err := c.CopyFile("file::testfile", "testfile2")
 	FailOnError(t, err)
 }
 
 func TestRenameFile(t *testing.T) {
-	c := createStubHTTPClient(t, "rename-file-req.xml", "rename-file-res.xml")
+	c := createStubHTTPClientSingle(t, "rename-file")
 	err := c.RenameFile("file::testfile2", "testfile1")
 	FailOnError(t, err)
 }
 
 func TestGetFileHandle(t *testing.T) {
-	c := createStubHTTPClient(t, "get-file-handle-req.xml", "get-file-handle-res.xml")
+	c := createStubHTTPClientSingle(t, "get-file-handle")
 	res, err := c.GetFileHandle("testfile")
 	FailOnError(t, err)
 
@@ -126,7 +126,7 @@ func TestGetFileHandle(t *testing.T) {
 }
 
 func TestGetFileAsRendered(t *testing.T) {
-	c := createStubHTTPClient(t, "get-file-as-rendered-req.xml", "get-file-as-rendered-res.xml")
+	c := createStubHTTPClientSingle(t, "get-file-as-rendered")
 	res, err := c.GetFileAsRendered("testfile")
 	FailOnError(t, err)
 
