@@ -22,7 +22,7 @@ import (
 )
 
 func TestGetProfiles(t *testing.T) {
-	c := createStubHTTPClient(t, "get-profiles-req.xml", "get-profiles-res.xml")
+	c := createStubHTTPClientSingle(t, "get-profiles")
 	profiles, err := c.GetProfiles()
 	FailOnError(t, err)
 
@@ -32,7 +32,7 @@ func TestGetProfiles(t *testing.T) {
 }
 
 func TestGetProfile(t *testing.T) {
-	c := createStubHTTPClient(t, "get-profile-req.xml", "get-profile-res.xml")
+	c := createStubHTTPClientSingle(t, "get-profile")
 	profile, err := c.GetProfile("Ubuntu-20.04-x86_64")
 	FailOnError(t, err)
 
@@ -42,19 +42,19 @@ func TestGetProfile(t *testing.T) {
 }
 
 func TestDeleteProfile(t *testing.T) {
-	c := createStubHTTPClient(t, "delete-profile-req.xml", "delete-profile-res.xml")
+	c := createStubHTTPClientSingle(t, "delete-profile")
 	err := c.DeleteProfile("test")
 	FailOnError(t, err)
 }
 
 func TestDeleteProfileRecursive(t *testing.T) {
-	c := createStubHTTPClient(t, "delete-profile-req.xml", "delete-profile-res.xml")
+	c := createStubHTTPClientSingle(t, "delete-profile")
 	err := c.DeleteProfileRecursive("test", false)
 	FailOnError(t, err)
 }
 
 func TestListProfileNames(t *testing.T) {
-	c := createStubHTTPClient(t, "get-item-names-profile-req.xml", "get-item-names-profile-res.xml")
+	c := createStubHTTPClientSingle(t, "get-item-names-profile")
 	profiles, err := c.ListProfileNames()
 	FailOnError(t, err)
 
@@ -64,7 +64,7 @@ func TestListProfileNames(t *testing.T) {
 }
 
 func TestGetProfilesSince(t *testing.T) {
-	c := createStubHTTPClient(t, "get-profiles-since-req.xml", "get-profiles-since-res.xml")
+	c := createStubHTTPClientSingle(t, "get-profiles-since")
 	profiles, err := c.GetProfilesSince(time.Date(1970, 1, 1, 0, 0, 0, 0, time.UTC))
 	FailOnError(t, err)
 
@@ -74,7 +74,7 @@ func TestGetProfilesSince(t *testing.T) {
 }
 
 func TestFindProfile(t *testing.T) {
-	c := createStubHTTPClient(t, "find-profile-req.xml", "find-profile-res.xml")
+	c := createStubHTTPClientSingle(t, "find-profile")
 	criteria := make(map[string]interface{}, 1)
 	criteria["name"] = "test"
 	profiles, err := c.FindProfile(criteria)
@@ -86,7 +86,7 @@ func TestFindProfile(t *testing.T) {
 }
 
 func TestFindProfileNames(t *testing.T) {
-	c := createStubHTTPClient(t, "find-profile-names-req.xml", "find-profile-names-res.xml")
+	c := createStubHTTPClientSingle(t, "find-profile-names")
 	criteria := make(map[string]interface{}, 1)
 	criteria["name"] = "test"
 	profiles, err := c.FindProfileNames(criteria)
@@ -98,25 +98,25 @@ func TestFindProfileNames(t *testing.T) {
 }
 
 func TestSaveProfile(t *testing.T) {
-	c := createStubHTTPClient(t, "save-profile-req.xml", "save-profile-res.xml")
+	c := createStubHTTPClientSingle(t, "save-profile")
 	err := c.SaveProfile("profile::testprof", "bypass")
 	FailOnError(t, err)
 }
 
 func TestCopyProfile(t *testing.T) {
-	c := createStubHTTPClient(t, "copy-profile-req.xml", "copy-profile-res.xml")
+	c := createStubHTTPClientSingle(t, "copy-profile")
 	err := c.CopyProfile("profile::testprof", "testprof2")
 	FailOnError(t, err)
 }
 
 func TestRenameProfile(t *testing.T) {
-	c := createStubHTTPClient(t, "rename-profile-req.xml", "rename-profile-res.xml")
+	c := createStubHTTPClientSingle(t, "rename-profile")
 	err := c.RenameProfile("profile::testprof2", "testprof1")
 	FailOnError(t, err)
 }
 
 func TestGetProfileHandle(t *testing.T) {
-	c := createStubHTTPClient(t, "get-profile-handle-req.xml", "get-profile-handle-res.xml")
+	c := createStubHTTPClientSingle(t, "get-profile-handle")
 	res, err := c.GetProfileHandle("testprof")
 	FailOnError(t, err)
 
